backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
alembic==1.12.1
pydantic-settings==2.1.0
pydantic==2.10.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.10.0
python-engineio==4.8.0
redis==5.0.1
httpx==0.25.2
aiofiles==23.2.1
azure-storage-blob==12.18.3
pdfplumber==0.10.3
python-docx==1.0.1
chromadb==0.4.24
tiktoken==0.12.0
